Allow ssl certificate validation bypass.
Most people don't want to shell out the 10 or 15 bucks for an ssl cert just so people can't snoop on thier traffic. Until the EFF gets us free certs (https://www.eff.org/press/releases/new-free-certificate-authority-dramatically-increase-encrypted-internet-traffic) we will need a way to bypass server ssl certificate validation. This needs to be optional as there are some crazy people, such as myself, who actually purchased a certificate. This change needs to be vetted on Windows 8 and phone. I'm not 100% sure this API is supported and I don't have devices to test against.

This is incomplete. I'm still looking at the best way to expose this via the API. Does adding a property in connectionmanager seem appropiate?